Move tests and fix console logs
