+++ tests/unit/static-build/all.ts
@@ -1,0 +1,2 @@
+import './getFeatures';
+import './StaticOptimizePlugin';
